refactor TransactionSummary
